refactor structire so settingsverifier in settings util
