+++ util/session/sessionmanager.go
@@ -2,8 +2,6 @@ package session
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/base64"
 	"fmt"
 	"net"
 	"net/http"
@@ -122,18 +120,6 @@ func (mgr *SessionManager) Parse(tokenString string) (jwt.Claims, error) {
 		return nil, err
 	}
 	return token.Claims, nil
-}
-
-// MakeSignature generates a cryptographically-secure pseudo-random token, based on a given number of random bytes, for signing purposes.
-func MakeSignature(size int) ([]byte, error) {
-	b := make([]byte, size)
-	_, err := rand.Read(b)
-	if err != nil {
-		b = nil
-	}
-	// base64 encode it so signing key can be typed into validation utilities
-	b = []byte(base64.StdEncoding.EncodeToString(b))
-	return b, err
 }
 
 // VerifyUsernamePassword verifies if a username/password combo is correct
