+++ util/util.go
@@ -1,6 +1,8 @@
 package util
 
 import (
+	"crypto/rand"
+	"encoding/base64"
 	"time"
 )
 
@@ -36,4 +38,16 @@ func Wait(timeout uint, f func(chan<- bool)) bool {
 			return false
 		}
 	}
+}
+
+// MakeSignature generates a cryptographically-secure pseudo-random token, based on a given number of random bytes, for signing purposes.
+func MakeSignature(size int) ([]byte, error) {
+	b := make([]byte, size)
+	_, err := rand.Read(b)
+	if err != nil {
+		b = nil
+	}
+	// base64 encode it so signing key can be typed into validation utilities
+	b = []byte(base64.StdEncoding.EncodeToString(b))
+	return b, err
 }
